“add_descriptions”
