+++ src/homework1/10_line.py
@@ -1,3 +1,4 @@
+"""10 line"""
 from time import localtime
 
 activities = {8: 'Sleeping',
