+++ src/homework1/11_line.py
@@ -1,3 +1,4 @@
+"""11 line"""
 REFRAIN = '''
 %d bottles of beer on the wall,
 %d bottles of beer,
