+++ src/homework1/12_line.py
@@ -1,3 +1,6 @@
+"""12 line"""
+
+
 class BankAccount(object):
     def __init__(self, initial_balance=0):
         self.balance = initial_balance
