+++ src/homework1/13_line.py
@@ -1,12 +1,8 @@
+"""13 line"""
+import unittest
+
+
 def median(pool):
-
-    """Statistical median to demonstrate doctest.
-
-    >>> median([2, 9, 9, 7, 9, 2, 4, 5, 8])
-
-    6 #change to 7 in order to pass the test
-    """
-
     copy = sorted(pool)
     size = len(copy)
     if size % 2 == 1:
@@ -15,6 +11,10 @@ def median(pool):
         return (copy[int(size / 2 - 1)] + copy[int(size / 2)]) / 2
 
 
+class TestMedian(unittest.TestCase):
+    def testMedian(self):
+        self.assertEqual(median([2, 9, 9, 7, 9, 2, 4, 5, 8]), 7)
+
+
 if __name__ == '__main__':
-    import doctest
-    doctest.testmod()
+    unittest.main()
