+++ src/homework1/14_line.py
@@ -1,7 +1,12 @@
-import unittest
+"""14 line"""
 
 
 def median(pool):
+    """Statistical median to demonstrate doctest.
+    >>> median([2, 9, 9, 7, 9, 2, 4, 5, 8])
+    6 #change to 7 in order to pass the test
+    """
+
     copy = sorted(pool)
     size = len(copy)
     if size % 2 == 1:
@@ -10,10 +15,6 @@ def median(pool):
         return (copy[int(size / 2 - 1)] + copy[int(size / 2)]) / 2
 
 
-class TestMedian(unittest.TestCase):
-    def testMedian(self):
-        self.assertEqual(median([2, 9, 9, 7, 9, 2, 4, 5, 8]), 7)
-
-
 if __name__ == '__main__':
-    unittest.main()
+    import doctest
+    doctest.testmod()
