+++ src/homework1/15_line.py
@@ -1,3 +1,4 @@
+"""15 line"""
 from itertools import groupby
 lines = '''
 This is the
