+++ src/homework1/16_line.py
@@ -1,3 +1,4 @@
+"""16 line"""
 import csv
