+++ src/homework1/18_line.py
@@ -1,3 +1,4 @@
+"""18 line"""
 BOARD_SIZE = 8
