+++ src/homework1/1_line.py
@@ -1,1 +1,2 @@
+"""1 line"""
 print('Привет, мир!')
