+++ src/homework1/20_line.py
@@ -1,3 +1,4 @@
+"""20 line"""
 import itertools
