+++ src/homework1/21_line.py
@@ -1,3 +1,4 @@
+"""21 line"""
 # From http://effbot.org/zone/element-index.htm
 import xml.etree.ElementTree as eTree
