+++ src/homework1/2_line.py
@@ -1,2 +1,3 @@
+"""2 line"""
 name = input('What is your name?\n')
 print('Hi, %s.' % name)
