+++ src/homework1/3_line.py
@@ -1,3 +1,4 @@
+"""3 line"""
 friends = ['john', 'pat', 'gary', 'michael']
 for i, name in enumerate(friends):
     print("iteration {iteration} is {name}".format(iteration=i, name=name))
