+++ src/homework1/4_line.py
@@ -1,3 +1,4 @@
+"""4 line"""
 parents, babies = (1, 1)
 while babies < 100:
     print('This generation has {0} babies'.format(babies))
