+++ src/homework1/5_line.py
@@ -1,3 +1,6 @@
+"""5 line"""
+
+
 def greet(name):
     print('Hello', name)
