+++ src/homework1/6_line.py
@@ -1,3 +1,4 @@
+"""6 line"""
 import re
 for test_string in ['555-1212', 'ILL-EGAL']:
     if re.match(r'^\d{3}-\d{4}$', test_string):
