+++ src/homework1/7_line.py
@@ -1,3 +1,4 @@
+"""7 line"""
 prices = {'apple': 0.40, 'banana': 0.50}
 my_purchase = {
     'apple': 1,
