+++ src/homework1/8_line.py
@@ -1,3 +1,4 @@
+"""8 line"""
 # This program adds up integers that have been passed as arguments in the command line
 import sys
 try:
